+++ opencv/profile.py
@@ -134,8 +134,10 @@ def create_image_section(inputfile, ext, host, fileid, key):
                 logger.debug("[%s] created section and previews of type %s", fileid, ext)
 
     finally:
-        #os.remove(previewfile)     
-        os.remove(sectionfile)  
+        #os.remove(previewfile)
+        if os.path.isfile(sectionfile):     
+            os.remove(sectionfile)     
+       
         
 
 def get_image_data(imagefile):
@@ -236,7 +238,13 @@ def on_message(channel, method, header, body):
                             body=json.dumps(statusreport))
 
         if inputfile is not None:
-            os.remove(inputfile)
+            try:
+                os.remove(inputfile)
+            except OSError:
+                pass
+            except UnboundLocalError:
+                pass  
+            
 
 
 if __name__ == "__main__":
